+++ foyer_oualhieya/src/main/java/tn/esprit/foyer_oualhieya/DAO/Entities/Chambre.java
@@ -18,7 +18,7 @@ public class Chambre {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "identifiantChambre")
-    private long idChambre;
+    private int idChambre;
     private long numeroChambre;
     private TypeChambre typeC;
